requirements.txt
# Streamlit
streamlit==1.39.0

# LangChain Core
langchain-core==0.3.15

# LangChain Modules
langchain-openai==0.2.5
langchain-ollama==0.2.0
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.2
langchain-community==0.3.5

# Additional Libraries
torch==2.3.1
faiss-cpu==1.9.0

# dotenv for environment variable loading
python-dotenv==1.0.1

# Outras Dependências Essenciais
accelerate==1.1.0
aiohttp==3.10.10
altair==5.4.1
annotated-types==0.7.0
attrs==24.2.0
bcrypt==4.2.0
bitsandbytes==0.44.1
dataclasses-json==0.6.7
Deprecated==1.2.14
einops==0.8.0
fsspec==2024.10.0
google-auth==2.35.0
huggingface-hub==0.26.2
jsonschema==4.23.0
langsmith==0.1.139
pandas==2.2.3
protobuf==4.25.5
pypdf==5.1.0
python-dateutil==2.9.0.post0
pytz==2022.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
SQLAlchemy==2.0.35
tqdm==4.66.6
transformers==4.46.1
tzdata==2024.2
uvicorn==0.32.0
websocket-client==1.8.0
yarl==1.17.1
wikipedia-api
requests newspaper3k
